extracting-domain-name-information/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
future==0.18.3
idna==3.4
python-whois==0.8.0
requests==2.28.2
urllib3==1.26.14
